+++ MyExperiments/Random.h
@@ -8,9 +8,9 @@ public:
 	explicit Random(const unsigned long long min = 0, const unsigned long long max = 100);
 
 	auto get_random(void) -> unsigned long long;
-	auto set_min(const int min) -> void;
-	auto set_max(const int max) -> void;
-	auto set_min_max(const int min, int max) -> void;
+	auto set_min(const unsigned long long min) -> void;
+	auto set_max(const unsigned long long max) -> void;
+	auto set_min_max(const unsigned long long min, unsigned long long max) -> void;
 
 private:
 	unsigned long long min_;
